chore: remove legacy planner hook
