+++ src/components/planner/PlannerPage.tsx
@@ -16,7 +16,8 @@ import WeekNotes from "./WeekNotes";
 import WeekPicker from "./WeekPicker";
 import DayRow from "./DayRow";
 import ScrollTopFloatingButton from "./ScrollTopFloatingButton";
-import { useFocusDate, useWeek, type ISODate } from "./usePlanner";
+import { useFocusDate, useWeek } from "./useFocusDate";
+import type { ISODate } from "./plannerStore";
 import { PlannerProvider } from "./plannerStore";
 
 /* ───────── Page body under provider ───────── */
