+++ src/components/planner/TodayHero.tsx
@@ -9,7 +9,9 @@
 import { useMemo, useRef, useState, useEffect } from "react";
 import { cn } from "@/lib/utils";
 import { toISODate } from "@/lib/date";
-import { useFocusDate, useSelectedProject, useSelectedTask, type ISODate } from "./usePlanner";
+import { useFocusDate } from "./useFocusDate";
+import { useSelectedProject, useSelectedTask } from "./useSelection";
+import type { ISODate } from "./plannerStore";
 import { useDay } from "./useDay";
 import CheckCircle from "@/components/ui/toggles/CheckCircle";
 import Input from "@/components/ui/primitives/Input";
